utils/requirement.txt
# Note: only unusual dependencies are listed below

pytorch
torchvision
tqdm
lmdb